Refactor files into duke.management package.
