+++ src/main/java/duke/MainWindow.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.management.Ui;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
