+++ src/main/java/duke/management/Parser.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.management;
 
+import duke.DukeException;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
