+++ src/main/java/duke/management/TaskList.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.management;
 
+import duke.DukeException;
 import duke.task.Task;
 
 import java.util.ArrayList;
